chore(file-uploader): update file-uploader to sb v7
