+++ packages/carbon-web-components/src/components/file-uploader/file-uploader.mdx
@@ -1,5 +1,8 @@
-import { Props, Description } from '@storybook/addon-docs/blocks';
+import { ArgsTable, Meta, Description } from '@storybook/addon-docs/blocks';
 import { cdnJs, cdnCss } from '../../globals/internal/storybook-cdn';
+import * as FileUploaderStories from './file-uploader.stories';
+
+<Meta of={FileUploaderStories} />
 
 # File uploader
 
@@ -40,15 +43,15 @@ import '@carbon/web-components/es/components/file-uploader/index.js';
 
 ## `<cds-file-uploader>` attributes, properties and events
 
-<Props of="cds-file-uploader" />
+<ArgsTable of="cds-file-uploader" />
 
 ## `<cds-file-uploader-drop-container>` attributes, properties and events
 
-<Props of="cds-file-uploader-drop-container" />
+<ArgsTable of="cds-file-uploader-drop-container" />
 
 ## `<cds-file-uploader-button>` attributes, properties and events
 
-<Props of="cds-file-uploader-button" />
+<ArgsTable of="cds-file-uploader-button" />
 
 ## `<cds-file-uploader-item>` attributes, properties and events
 
@@ -56,4 +59,4 @@ Note: For `boolean` attributes, `true` means simply setting the attribute (e.g.
 `<cds-file-uploader-item invalid>`) and `false` means not setting the attribute
 (e.g. `<cds-file-uploader-item>` without `invalid` attribute).
 
-<Props of="cds-file-uploader-item" />
+<ArgsTable of="cds-file-uploader-item" />
